refactor: move pql_reference module into the tools directory and update imports
