+++ plane_mcp/tools/modules.py
@@ -15,11 +15,10 @@ from plane.models.modules import (
     UpdateModule,
 )
 from plane.models.query_params import WorkItemQueryParams
-from plane.models.work_items import WorkItem
 from pydantic import Field
 
 from plane_mcp.client import get_plane_client_context
-from plane_mcp.pql_reference import PQL_FIELD_HINT, PQL_FULL_REFERENCE
+from plane_mcp.tools.pql_reference import PQL_FIELD_HINT, PQL_FULL_REFERENCE
 
 logger = get_logger(__name__)
